Feature/fix global state (#19)
This is a long overdue fix for the global state of the application. Currently, we are using electron-shared-store for the state, which also stores the controllers, which really doesn't make sense. The flow now is:

Page calls a controller -> controller sends an ipc message to main -> main dispatches the message to the proper service function, which utilizes a loki service created in main -> the service returns a result to the message -> the message gets returned to the controller, and then the page that called it. This allows for a central store for a loki service. This is being built with the consideration that we can have multiple loki services open at a time (in the future).