Syncs time spent with Trello card description
Ensures that the time spent on a node is synchronized with the Trello card description, using a `banflow:timeSpent` tag.

This change introduces helper functions to extract and set the `banflow:timeSpent` value in the description, maintaining consistency between the application and Trello. When the time spent is updated, the Trello description is updated accordingly. It also parses the description when creating a node from Trello data, extracting the time spent value.